pyproject.toml
[tool.poetry]
name = "multi-tenant-test"
version = "0.1.0"
description = "A multi-tenant application with FastAPI backend and Streamlit frontend, integrating OpenAI LLMs."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/your-repo/multi-tenant-test"
repository = "https://github.com/your-repo/multi-tenant-test"
keywords = ["fastapi", "streamlit", "multi-tenancy", "openai", "llm"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.95.2"
uvicorn = "0.22.0"
pydantic = "1.10.7"
openai = "0.27.8"
sqlalchemy = "1.4.46"
passlib = {extras = ["bcrypt"], version = "1.7.4"}
python-jose = "3.3.0"
streamlit = "1.24.0"
requests = "2.31.0"
python-dotenv = "1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.20.3"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"